refactor: extract bp_channels to routes/channels.py
Phase 5.7 of incremental modularisation. Moves the 21 handlers
registered on bp_channels (Telegram, iMessage, WhatsApp, Signal,
Discord, Slack, IRC, WebChat, Google Chat, BlueBubbles, MS Teams,
Matrix, Mattermost, LINE, Nostr, Twitch, Feishu, Zalo, Tlon, Synology
Chat, Nextcloud Talk) into routes/channels.py following the existing
late-import-of-dashboard pattern.

Module-level helpers (_get_log_dirs, _grep_log_file, and the shared
_generic_channel_data fetcher) stay in dashboard.py and are reached
via `import dashboard as _d`. Pure mechanical move — zero behaviour
change.

dashboard.py 30062 -> 28603 (-1459 lines). routes/channels.py is
1505 lines. All 21 channel endpoints + cross-phase smoke tests
(/api/overview, /api/sessions, /api/usage, /api/crons, /api/health,
/api/brain-history, /api/alerts/rules, /api/budget/status) return 200.
